requirements.txt
python-decouple
aiogram
requests
bs4